Switch mouse/gamepad with baton's getActiveDevice
Instead of making it static, use getActiveDevice to determine what the
player is inputting.

It uses the deadzone to ignore small gamepad
inputs, which is fine but requiring a button would be better. Oh well.
Requires keyboard input to switch user to the mouse (probably because no
mouse inputs are bound).