Prevent mouse aiming&shooting from losing focus
Since we're not fullscreen, don't want to accidental clicking on
background apps while frantically aiming.

On macOS, this looks bad (jitters along border for me) and doesn't work
awesomely (sometimes you can still click on background apps). Not sure
if it's better on other platforms, but it's better than nothing.